requirements.txt
pandas
numpy
matplotlib
scikit-learn
streamlit
kafka-python
psycopg2-binary
mne
